feat(events): adding create, read, delete feature

+++ src/main/java/com/eventmanagement/EventManagementBackend/infrastructure/users/repository/UsersRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsersRepository extends JpaRepository<UsersAccount, Long> {
+public interface UsersRepository extends JpaRepository<UsersAccount, Integer> {
 }
